feat(sst): add RawFrame structure and related functions for raw stack frame resolution
This change introduces a new `RawFrame` structure to represent raw stack
frames,
allowing for the resolution of addresses to their corresponding module
and
offset information. It also adds functions to resolve a single address
or
batch of addresses to raw frame information, enhancing the stack trace
capabilities of the library. This is done to provide more granular
information about stack frames, which can be useful for debugging and
analysis in using `addr2line`.